chore: rename element -> component
